# server/internal/handlers/openapi.yaml
openapi: "3.1.0"
info:
  title: Secsy PKI
  description: >-
    HSM-backed enterprise PKI and secret-encryption platform. Exposes X.509
    certificate-authority management (root/intermediate setup, issuance, renewal,
    revocation with CRL/OCSP), SSH certificate signing, HSM-backed envelope
    encryption for secrets, certificate-expiry monitoring, RBAC/audit
    administration, ACME (RFC 8555) directory, and operational endpoints. Every
    private-key operation runs inside the configured key provider (PKCS#11 HSM /
    YubiHSM); key material never leaves the token.


    Protocol parity: the core certificate-lifecycle operations documented here —
    issue (from CSR/template), renew, revoke, get certificate/status, list, and
    CRL/OCSP metadata — are additionally exposed over a gRPC API (PKIService,
    proto/pki/v1/pki.proto) with server reflection, enforcing the same operator
    authentication (Bearer OIDC / Basic root / mutual-TLS), RBAC, tenant scoping,
    and tamper-evident audit as these REST endpoints. See docs/grpc-api.md.
  version: "2.0.0"
  license:
    name: MIT
  contact:
    name: Secsy PKI
    url: https://github.com/blechschmidt/secsy-pki
servers:
  - url: /
    description: The Secsy PKI server root (all paths are absolute).
security:
  - basicAuth: []
  - bearerAuth: []
...
paths:
...
  /api/tenants/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: string }
        description: Tenant id.
    get:
      operationId: getTenant
      tags: [Tenants]
      summary: Get a tenant by id
      description: Readable by platform operators and by members of the tenant.
      responses:
        "200":
          description: The requested tenant.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Tenant" }
        "404": { $ref: "#/components/responses/NotFound" }
    put:
      operationId: updateTenant
      tags: [Tenants]
      summary: Update a tenant's name, KEK label, or quotas (platform admin only)
      description: >-
        Partial update; absent fields are unchanged. Quota changes take effect
        on the next issuance/secret operation — enforcement is fail-closed on
        the issuance (REST/ACME/SCEP/EST/CMP/gRPC/SSH) and secret paths, where
        exhaustion answers 429 with code quota_exceeded.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdateTenantRequest" }
      responses:
        "200":
          description: The updated tenant.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Tenant" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
    delete:
      operationId: deleteTenant
      tags: [Tenants]
      summary: Delete an empty tenant (platform admin only)
      description: >-
        Refused for the default tenant and for any tenant that still owns CAs.
      responses:
        "200":
          description: The tenant was deleted.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/StatusResponse" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "409": { $ref: "#/components/responses/BadRequest" }
...
  /api/secret/encrypt:
    post:
      operationId: encryptSecret
      tags: [Secrets]
      summary: Envelope-encrypt a secret
      parameters:
        - $ref: "#/components/parameters/TenantHeader"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/EncryptRequest" }
      responses:
        "200":
          description: The versioned envelope.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/EncryptResponse" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "413":
          description: Plaintext exceeds the maximum size.
  /api/secret/decrypt:
    post:
      operationId: decryptSecret
      tags: [Secrets]
      summary: Decrypt an envelope
      parameters:
        - $ref: "#/components/parameters/TenantHeader"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/DecryptRequest" }
      responses:
        "200":
          description: The recovered plaintext (base64).
          content:
            application/json:
              schema: { $ref: "#/components/schemas/DecryptResponse" }
        "400":
          description: Decryption failed (generic, oracle-resistant).
        "403": { $ref: "#/components/responses/Forbidden" }

  # ---------------------------------------------------------------------------
  # HSM
  # ---------------------------------------------------------------------------
...
components:
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic
      description: Built-in root user (username/password).
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: OIDC access token.
  parameters:
...
    TenantHeader:
      name: X-Secsy-Tenant
      in: header
      required: false
      description: >-
        Selects the tenant (id or slug) whose secret KEK seals/opens the
        envelope. Omit to use the default tenant and the deployment-wide KEK.
      schema: { type: string }
...
  responses:
    BadRequest:
      description: The request was malformed.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
    Forbidden:
      description: The caller lacks the required permission.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    NotFound:
      description: The resource was not found.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error: { type: string }
    StatusResponse:
      type: object
      properties:
        status: { type: string }
...
    Tenant:
      type: object
      properties:
        id: { type: string }
        slug: { type: string, description: URL/CLI-friendly unique identifier. }
        name: { type: string }
        status: { type: string, enum: [active, suspended] }
        kek_label: { type: string, description: Optional per-tenant secret KEK. }
        created_at: { type: string, format: date-time }
        quotas: { $ref: "#/components/schemas/TenantQuotas" }
    TenantQuotas:
      type: object
      description: >-
        Per-tenant consumption ceilings. Zero (or omitted) means unlimited for
        that dimension. Daily quotas meter per UTC day and reset at UTC
        midnight; exhaustion is reported as HTTP 429 with code quota_exceeded
        and, for daily windows, a Retry-After header.
      properties:
        max_certs_per_day:
          type: integer
          format: int64
          description: Max certificates issued per UTC day (X.509 and SSH), across all the tenant's CAs and every enrollment protocol.
        max_active_certs:
          type: integer
          format: int64
          description: Ceiling on unexpired, unrevoked X.509 certificates; revocation or expiry frees room.
        max_secret_ops_per_day:
          type: integer
          format: int64
          description: Max envelope encrypt/decrypt operations per UTC day.
        rate_limit_per_second:
          type: number
          format: double
          description: Per-tenant request rate override for public enrollment endpoints (0 inherits rate_limit.per_tenant).
        rate_limit_burst:
          type: number
          format: double
          description: Burst capacity accompanying rate_limit_per_second.
...
    UpdateTenantRequest:
      type: object
      description: >-
        Partial update: absent fields are left unchanged. quotas, when present,
        replaces the whole quota set.
      properties:
        name: { type: string }
        kek_label: { type: string }
        quotas: { $ref: "#/components/schemas/TenantQuotas" }
...
    EncryptRequest:
      type: object
      required: [plaintext]
      properties:
        plaintext: { type: string, description: base64-encoded plaintext }
        context: { type: string, description: base64-encoded optional AAD }
        escrow:
          type: boolean
          description: >-
            Additionally escrow the data key to the configured M-of-N recovery
            agents so it can be recovered under dual control. Requires
            secret.escrow to be configured; recovery itself is a CLI-only
            ceremony (secsy-secret recover).
    EncryptResponse:
      type: object
      properties:
        envelope:
          type: object
          additionalProperties: true
          description: Opaque versioned envelope document.
    DecryptRequest:
      type: object
      required: [envelope]
      properties:
        envelope:
          type: object
          additionalProperties: true
        context: { type: string, description: base64-encoded optional AAD }
    DecryptResponse:
      type: object
      properties:
        plaintext: { type: string, description: base64-encoded plaintext }
...